brackup-target: Optionally use local metafiles instead of possibly encrypted remote ones
This patch is based on one by Shaun Amott <shaun@inerd.com> from 2008
[1].  I modified it to apply to optionally take a directory to use
instead of always using the cwd.  I also cleaned up the code for more
error handling/debugging, style choices, and simplicity (globs instead
of opendir/readdir).  It needed a little massaging to apply to the
current version of Brackup and conform to new default filenames.

There are two major caveats with this approach, currently noted in the
option documentation.

[1] http://lists.danga.com/pipermail/brackup/2008-October/000178.html
    http://www.inerd.com/patches/brackup-210-gc-local-metafiles.diff